requirements.txt
google-generativeai
feedparser
Pillow
requests
python-dotenv
